requirements.txt
Django==2.2.6
django-picklefield==2.0
djangorestframework==3.10.3
jsonfield==2.0.2
mysqlclient==1.4.5
pickleshare==0.7.5
requests==2.22.0
retry==0.9.2
six==1.12.0
supervisor==4.1.0
